Create TemplateExecutor and make StringExec tests pass
