+++ execFileSystem.go
@@ -4,6 +4,7 @@ import (
 	"bytes"
 	"fmt"
 	"html/template"
+	"io"
 	"net/http"
 )
 
@@ -95,4 +96,14 @@ func (fe *FileSystemExecutor) constructTemplate(view *View) (*template.Template,
 		}
 	}
 	return out, nil
+}
+
+// readStringAndClose ensures that the supplied read closer is closed
+func readStringAndClose(buffer *bytes.Buffer, rc io.ReadCloser) (string, error) {
+	defer rc.Close()
+	_, err := buffer.ReadFrom(rc)
+	if err != nil {
+		return "", err
+	}
+	return buffer.String(), nil
 }
